% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.r
\name{majority_label}
\alias{majority_label}
\title{Question 12: Return the Most Frequent Value (Majority Label)}
\usage{
majority_label(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of any type (character, numeric, factor, etc.).}

\item{na.rm}{Logical; if \code{TRUE}, \code{NA} values are removed before calculation. Default is \code{TRUE}.}
}
\value{
The most frequent value in the vector. If all values are \code{NA} or input is empty, returns \code{NA}.
}
\description{
Returns the most frequently occurring value in a vector, optionally excluding \code{NA}.
}
\details{
If there is a tie for the most frequent value, the function returns the first one encountered in the mode.
}
\examples{
majority_label(c("apple", "banana", "apple"))      # "apple"
majority_label(c(1, 1, 2, 2, 2, NA), na.rm = TRUE)  # 2
majority_label(c(NA, NA))                          # NA

}
